requirements.txt
arcadepy>=1.4.0
python-dotenv>=1.1.0